Drop dom and http extensions
these extensions provide no value

HTTP:
1. XHR and fetch are two very distinct APIs and trying to create a 3rd
on top of them to proxy to whichever one is supported by the browser
underneath is not easy, plus it’s not effective and may hinder than
help.
2. It’s much better to let the library user choose what he wants and
just put that a small mask on top of it.

DOM:
- This abstraction has no value at all, it’s just proxying into jQuery.
- Nowadays people are using web component libraries like React and Riot
and Polymer and almost never user jQuery for DOM manipulation. So yet
again it hinders and restricts more than helps

Close #4