Only lint the lint file and example code
1. Why is this change necessary?
Makes no sense to be linting the small modules before they have been
merged into a single file.

For instance prior to this change I would get an error for each file
due to the lack of 'use strict'; But I have the 'use strict' in wrapper
file, and thus should not need in each of the small module files.
One should suffice.