Have TravisCI do the build, lint and test
1. Why is this change necessary?

Currently I have to push a new f.js build for Travis to try out.
That means I can builds in dist/ that don't work and as a distribution
folder that seems wrong.

2. How does it address the issue?

It makes travis do the build itself and test based on what it built,
such that whenever I push new code into src, it will do the assemblying
and try out. And I can push only clean builds to dist