Finally understood noConflict :]
noConflict reverts the original value of the global namespace variable
we set when loading the library, i.e. there could be already some
value in window.F before the library is loaded, thus when we load it
we take the original value of window.F and stash it in a private
variable previousF. Then when we call noConflict it puts back the cached
value into the window.F.

The above means our lib is no longer accessible via window.F so thats
why in the end of no noConflict we return it, as to provide the caller
with the opportunity to put it in a new variable.